requirements.txt
# requirements.txt
pandas
numpy
requests
pybaseball
mlbstatsapi
python-telegram-bot